Updated calculate.py and functionslist
Added code to the following functions in calculate.py:
- calculate_mana_cost
- calculate_damage_per_second
- calculate_crit_chance

Also added calls to the functions to test the in-game values. This will
be added to unittests later. The unittests will be updated on a regular
basis with the actual values to be able to compare them to the actual
values.