Add init function and function to get node IDs from URL to passives.py
Added an init function to passives.py that sets the build URL that the
user has entered. Also added a function to base64 decode the build URL
and a function to retrieve the node IDs of the skilled nodes.

The function that does the base64 decoding returns a string, the
function that retrieves the node IDs of the skilled nodes returns a list
of integers.