Add function load_node_data to passives.py
Added the function load_node_data to the passives.py module. This
function currently only retrieve the node IDs of all the nodes in the
passive tree. However, later this function will retrieve all the
relevant information of the skill nodes, to support calculation of
bonuses from a given passive tree