Add function get_bonus_for_selected_nodes to passives.py
The function get_bonus_for_selected_nodes was added to the passives.py
module. The function currently retrieves the names and description of
all the skilled nodes and prints this information to the console.

Later this function will return a dictionary that can be used to extract
the passive bonuses per type.