Make widgets that need to be accesible by event handlers class-wide
A lot of the widgets in the GUI need to be accesible to event handler
functions that either need to retrieve the value of a widget, or alter
the value of the widget.

The system that PyQt4 uses to connect a function to a widget does not
allow for passing in arguments. Therefore, the widgets that might be
altered need to be class-wide accessible.