Add code to passives.py to calculate passive bonuses
Added code to passive.py to calculate the actual passive bonuses with
their values for a given passive tree URL. Some nodes are currently not
parsed properly, but this is something that's being looked into.

Over the next week, all the skill nodes of the tree will be checked
(actually work in progress at the time of this commit), and testing will
be done with decoding of bitcodes.