Create unique dropdowns/level boxes for the support skills
Because how how PyQt4 event handling works, each dropdown menu for the
support skills and the levels needs to have a unique name. This way we
can identify which dropdown menu has changed, and in the future this can
be used to properly update the values displayed in the other dropdown
menus.